Added Partial Snapshot Acquisition
We observed that the oracle could become the bottleneck as the number of clients increases. This is because each snapshot is a vector proportional to the number of clients. Before, transactions read this vector in the beginning. With the new idea, a transaction only reads the part of the entire snapshot that belongs to its readset, which renders in MUCH SMALLER snapshots